scoreboard/requirements.txt
pyhamcrest==2.0.2
pytest==6.1.1
pytest_watch==4.2.0
click==8.1.3